got rid of JSON-centric testing version, moving towards more adaptable module
